Rewrite Pc

+++ src/consts.rs
@@ -1,7 +1,7 @@
 //! Shared constants of program space.
 
 /// Width of program space.
-pub const WIDTH: i16 = 80;
+pub const WIDTH: usize = 80;
 
 /// Height of program space.
-pub const HEIGHT: i16 = 25;
+pub const HEIGHT: usize = 25;
